+++ gradle/libs.versions.toml
@@ -7,12 +7,10 @@ gratatouille = "0.0.10"
 json = "org.jetbrains.kotlinx:kotlinx-serialization-json:1.8.1"
 okio = "com.squareup.okio:okio:3.8.0"
 okhttp = "com.squareup.okhttp3:okhttp:4.12.0"
-okhttp-logging-interceptor = "com.squareup.okhttp3:logging-interceptor:4.12.0"
 mockwebserver = "com.squareup.okhttp3:mockwebserver:4.12.0"
 gradle-min = "dev.gradleplugins:gradle-api:8.8"
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test" }
 xmlutil = "io.github.pdvrieze.xmlutil:serialization:0.91.1"
-
 
 [plugins]
 kgp = { id = "org.jetbrains.kotlin.jvm", version.ref = "kgp" }
